+++ test_output/config/config.go
@@ -1,0 +1,3 @@
+package config
+const ConfigHost = "localhost"
+const ConfigPort = 8080