+++ test_output/database/database.go
@@ -1,0 +1,3 @@
+package database
+const DatabaseHost = "localhost"
+const DatabasePort = 5432